added Direction, and move WASD keys
