+++ src/main/java/ru/mipt/bit/platformer/keys/Key.java
@@ -1,7 +1,5 @@
 package ru.mipt.bit.platformer.keys;
 
-import com.badlogic.gdx.math.GridPoint2;
-
 public interface Key {
     default boolean isPressed(){
         return false;
