+++ src/main/java/ru/mipt/bit/platformer/keys/MovementKey.java
@@ -1,35 +1,10 @@
 package ru.mipt.bit.platformer.keys;
 
 import com.badlogic.gdx.Gdx;
-import com.badlogic.gdx.math.GridPoint2;
 import ru.mipt.bit.platformer.objects.Movable;
 import ru.mipt.bit.platformer.objects.Object;
 
 import java.util.Arrays;
-
-import static ru.mipt.bit.platformer.util.GdxGameUtils.*;
-
-enum Direction{
-    UP   (new GridPoint2( 0,  1),   90f),
-    DOWN (new GridPoint2( 0, -1),  -90f),
-    LEFT (new GridPoint2(-1,  0), -180f),
-    RIGHT(new GridPoint2( 1,  0),    0f);
-    private final GridPoint2 direction;
-    private final float rotation;
-
-    Direction(GridPoint2 direction, float rotation) {
-        this.direction = direction;
-        this.rotation  = rotation;
-    }
-
-    public GridPoint2 getDirection() {
-        return direction;
-    }
-
-    public float getRotation() {
-        return rotation;
-    }
-}
 
 public abstract class MovementKey implements Key {
 
@@ -58,56 +33,9 @@ public abstract class MovementKey implements Key {
             if (!existCollisions(movable)) {
                 movable.changeDestinationCoordinates(dir.getDirection());
                 movable.setRotation(dir.getRotation());
+                movable.setMovementProgress(0f);
             }
         }
-    }
-}
-
-class UP extends MovementKey {
-
-    public UP(Object[] obstacles, Movable[] movables, int[] keys, Direction dir) {
-        super(obstacles, movables, keys, dir);
-    }
-
-    @Override
-    boolean existCollisions(Movable movable) {
-        return Arrays.stream(obstacles).anyMatch(obstacle -> obstacle.getCoordinates().equals(incrementedY(movable.getCoordinates())));
-    }
-}
-
-class LEFT extends MovementKey {
-
-    public LEFT(Object[] obstacles, Movable[] movables, int[] keys, Direction dir) {
-        super(obstacles, movables, keys, dir);
-    }
-
-    @Override
-    boolean existCollisions(Movable movable) {
-        return Arrays.stream(obstacles).anyMatch(obstacle -> obstacle.getCoordinates().equals(decrementedX(movable.getCoordinates())));
-    }
-}
-
-class DOWN extends MovementKey {
-
-    public DOWN(Object[] obstacles, Movable[] movables, int[] keys, Direction dir) {
-        super(obstacles, movables, keys, dir);
-    }
-
-    @Override
-    boolean existCollisions(Movable movable) {
-        return Arrays.stream(obstacles).anyMatch(obstacle -> obstacle.getCoordinates().equals(decrementedY(movable.getCoordinates())));
-    }
-}
-
-class RIGHT extends MovementKey {
-
-    public RIGHT(Object[] obstacles, Movable[] movables, int[] keys, Direction dir) {
-        super(obstacles, movables, keys, dir);
-    }
-
-    @Override
-    boolean existCollisions(Movable movable) {
-        return Arrays.stream(obstacles).anyMatch(obstacle -> obstacle.getCoordinates().equals(incrementedX(movable.getCoordinates())));
     }
 }
