+++ src/main/java/ru/mipt/bit/platformer/util/KeyPressHandler.java
@@ -5,9 +5,8 @@ import ru.mipt.bit.platformer.keys.Key;
 public abstract class KeyPressHandler {
     public static void handleKeyPress(Key ...keys) {
         for (Key k : keys) {
-            if (k.isPressed()) {
+            if (k.isPressed())
                 k.action();
-            }
         }
     }
 }
